google-ads-analysis/requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
requests>=2.28.0
python-dotenv>=1.0.0

# Google APIs
google-ads>=22.1.0
google-analytics-data>=0.18.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
google-generativeai>=0.3.0

# Data processing
beautifulsoup4>=4.11.0
Pillow>=9.0.0
lxml>=4.9.0

# Scheduling and automation
schedule>=1.2.0

# Utilities
rich>=13.0.0
typing-extensions>=4.0.0